API calls moved to Game.js
