+++ src/App.scss
@@ -76,6 +76,10 @@ main {
   flex-wrap: wrap;
 }
 
+.loadingText {
+  font-size: 2rem;
+}
+
 .wordBox {
   display: inline-block;
   border: 1px solid black;
